Clean up build problems, add a (lame) build script, add a simple sample
git-svn-id: svn://svn.code.sf.net/p/bfr/code/trunk@6 a57d90b3-4637-4882-8861-93cfad8ad01b

+++ BACnet/BACnetTask.cpp
@@ -96,7 +96,7 @@ const char* strtimeval(const timeval& t)
     static char buff[20]
     ;
     
-    sprintf( buff, "%d.%06d", t.tv_sec, t.tv_usec );
+    sprintf( buff, "%d.%06d", (int)t.tv_sec, (int)t.tv_usec );
     return buff;
 }
